backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask-cors
PyJWT
openai
bcrypt
python-dotenv
psycopg2-binary